# [PERF] stock: reduce number of queries when writing on smls