@robodoo 2024-12-12
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/190479.png)](https://mergebot.odoo.com/odoo/odoo/pull/190479)